pyproject.toml
[project]
name = "wn-llamaindex-agents-course"
version = "0.1.0"
description = "Hands-on LLM agents fundamentals with Llamaindex: agent concept, patterns, tools, need for framework and more"
authors = [
    {name = "Wojciech Nowak"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.1.1"
openai = "^1.99.9"
llama-index = "^0.14.8"
llama-index-utils-workflow = "^0.4.0"
llama-index-readers-wikipedia = "^0.4.1"
wikipedia = "^1.4.0"
datasets = "^4.4.1"
llama-index-readers-web = "^0.5.6"
ipywidgets = "^8.1.8"
llama-index-llms-llama-cpp = "^0.5.1"
llama-cpp-python = "^0.3.16"
llama-index-llms-huggingface = "^0.6.1"
llama-index-tools-wikipedia = "^0.4.1"
tavily-python = "^0.7.17"
duckdb = "^1.4.3"
matplotlib = "^3.10.8"
chromadb = "^1.4.0"
llama-index-vector-stores-chroma = "^0.5.5"
aiosqlite = "^0.22.1"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.5"
ipykernel = "^6.30.1"

[[tool.poetry.packages]]
include = "scripts"

[tool.poetry.scripts]
connectivity-check = "scripts.connectivity_check:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
